Moved class to Extensions
