+++ src/Hangfire.Raven/Storage/RavenStorageExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Hangfire.Raven.Extensions;
 using Raven.Client.Documents.Session;
 
 namespace Hangfire.Raven.Storage {
